refactor: Refactor: split orbit-mcp adapter.rs into focused submodules [ORB-00123]
Planned-By: claude
